test: consolidate provider capability coverage
